+++ src/main/scala-2.10/geostat/kriging/model/PoweredExponentialModel.scala
@@ -11,23 +11,24 @@ import scala.math._
  * @param p power
  *
  */
-class PoweredExponentialModel(val r: Double = 1.0, val s : Double = 1.0, val a : Double = 0.0, val p : Double = 1.0) extends Model {
+class PoweredExponentialModel(val r: Float = 1.0f, val s : Float = 1.0f, val a : Float = 0.0f, val p : Float = 1.0f) extends Model {
 
-  require(r > 0.0, "range must be greater than 0")
-  require(s > 0.0, "sill must be greater than 0")
-  require(a >= 0.0, "nugget must be greater than or equal 0")
+  require(r > 0.0f, "range must be greater than 0")
+  require(s > 0.0f, "sill must be greater than 0")
+  require(a >= 0.0f, "nugget must be greater than or equal 0")
   require(s >= a, "sill must be greater than nugget")
 
-  def variogram(h: Double) = {
-    var ret = 0.0
-    if (h > 0)
-      ret = a + (s - a) * (1.0 - exp(-3.0 * pow(h, p) / r))
+  def variogram(h: Float) = {
+    var ret = 0.0f
+    if (h > 0.0f)
+      ret = a + (s - a) * (1.0f - exp(-3.0f * pow(h, p) / r).toFloat)
     ret
   }
-  def covariogram(h: Double) = {
+  
+  def covariogram(h: Float) = {
     var ret = s
-    if (h > 0)
-      ret = (s - a) * exp(-3.0 * pow(h, p) / r)
+    if (h > 0.0f)
+      ret = (s - a) * exp(-3.0f * pow(h, p) / r).toFloat
     ret
   }
