+++ src/main/scala-2.10/geostat/kriging/model/WaveModel.scala
@@ -10,23 +10,24 @@ import scala.math._
  * @param a nugget
  *
  */
-class WaveModel(val w: Double = 1.0, val s : Double = 1.0, val a : Double = 0.0) extends Model {
+class WaveModel(val w: Float = 1.0f, val s : Float = 1.0f, val a : Float = 0.0f) extends Model {
 
-  require(w > 0.0, "wave intensity must be greater than 0")
-  require(s > 0.0, "sill must be greater than 0")
-  require(a >= 0.0, "nugget must be greater than or equal 0")
+  require(w > 0.0f, "wave intensity must be greater than 0")
+  require(s > 0.0f, "sill must be greater than 0")
+  require(a >= 0.0f, "nugget must be greater than or equal 0")
   require(s >= a, "sill must be greater than nugget")
 
-  def variogram(h: Double) = {
-    var ret = 0.0
-    if (h > 0)
-      ret = a + (s - a) * (1.0 - w * sin(h / w) / h)
+  def variogram(h: Float) = {
+    var ret = 0.0f
+    if (h > 0.0f)
+      ret = a + (s - a) * (1.0f - w * sin(h / w).toFloat / h)
     ret
   }
-  def covariogram(h: Double) = {
+  
+  def covariogram(h: Float) = {
     var ret = s
-    if (h > 0)
-      ret = (s - a) * (w * sin(h / w) / h)
+    if (h > 0.0f)
+      ret = (s - a) * (w * sin(h / w).toFloat / h)
     ret
   }
 }
